split into component the cornerbox and scanning line and remove some comments and console logs
